(fix): Add notifications (SignalR) and live-session features
Introduce real-time notifications and live-session enhancements across services. Adds a SignalR NotificationHub, INotificationPusher abstraction and NotificationPusher implementation, registers the hub and DI in AccountContentService, and updates messaging consumers to push notifications and log failures. Expose notification routes through API gateway (ocelot) and rename docker-compose project. Add GuestViewerTimeoutService to enforce 2-minute guest viewing limit and notify/disconnect expired guests, plus multiple LiveSessionService domain, infrastructure and migration changes (media, AzuraCast client, station/media mappings, repositories, and tests). Misc: update data seeder image URL and remove obsolete auth-service SQL migration files.

+++ services/live-session-service/LiveSessionService.Application/Features/Results/NowPlaying/StationNowPlayingResult.cs
@@ -27,10 +27,10 @@ public sealed class NowPlayingTrackResult
     public string? Genre { get; init; }
     public string? ArtUrl { get; init; }
     public string? Lyrics { get; init; }
-    public long PlayedAt { get; init; }
-    public long? Duration { get; init; }
-    public long Elapsed { get; init; }
-    public long Remaining { get; init; }
+    public double PlayedAt { get; init; }
+    public double? Duration { get; init; }
+    public double Elapsed { get; init; }
+    public double Remaining { get; init; }
     public bool IsRequest { get; init; }
 }
